rl_product_recommendation/requirements.txt
kafka_python==2.1.4
langchain==0.3.22
matplotlib==3.10.1
numpy==2.2.4
pandas==2.2.3
Pillow==11.2.0
protobuf==6.30.2
scipy==1.15.2
seaborn==0.13.2
streamlit==1.44.0
